+++ lib/src/page/dashboard/widget/item_response_widget.dart
@@ -56,6 +56,8 @@ class ItemResponseWidget extends StatelessWidget {
                 Expanded(
                   child: Text(
                     data.endpoint,
+                    maxLines: 1,
+                    overflow: TextOverflow.ellipsis,
                     style: TextStyle(
                       color: AppColor.primary,
                       fontWeight: FontWeight.bold,