+++ lib/src/page/resources/app_color.dart
@@ -3,4 +3,5 @@ import 'package:flutter/material.dart';
 abstract class AppColor {
   static Color primary = Colors.black.withOpacity(0.7);
   static Color white = Colors.white;
+  static Color red = Colors.red;
 }