+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 
+## 3.1.6
+
+* fix: copy image format
+* feat: add maxline in item list activity
+
+## 3.1.5
+
+* feat: add deprecated showFloating
+* feat: add copy detail activity
+* feat: add remove all activity 
+* feat: add copy to curl
+
 ## 3.1.4
 
 * style: fix UI for not using material 3